requirements.txt
asgiref==3.8.1
backgroundremover==0.3.4
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
commandlines==0.4.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
Django==5.2.3
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-resized==1.0.3
django-social-share==2.3.0
ffmpeg-python==0.2.0
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.3
fsspec==2025.5.1
future==1.0.0
hsh==1.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pil-plus==0.2
pillow==11.2.1
proglog==0.1.12
PyMatting==1.1.14
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
waitress==3.0.2
